+++ src/components/form/form-help/form-help.vue
@@ -1,5 +1,5 @@
 <template>
-	<span class="text-grey-500" data-test="form-help">
+	<span class="text-grey-500 dark:text-white/60" data-test="form-help">
 		<slot />
 	</span>
 </template>